+++ docs/changelog.rst
@@ -26,6 +26,12 @@ v0.20.0 (UNRELEASED)
 - Make the ``mopidy`` command print a friendly error message if the
   :mod:`gobject` Python module cannot be imported. (Fixes: :issue:`836`)
 
+**Configuration**
+
+- Add support for the log level value ``all`` to the loglevels configurations.
+  This can be used to show absolutely all log records, including those at
+  custom levels below ``DEBUG``.
+
 **Local backend**
 
 - Add cover URL to all scanned files with MusicBrainz album IDs. (Fixes:
@@ -50,7 +56,7 @@ v0.20.0 (UNRELEASED)
 
 - Enable browsing of artist references, in addition to albums and playlists.
   (PR: :issue:`884`)
-  
+
 - Share a single mapping between names and URIs across all MPD sessions. (Fixes:
   :issue:`934`, PR: :issue:`968`)